requirements.txt
pycryptodome
Flask[async]
requests
python-dotenv
pymongo
aiohttp
protobuf
